# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DebateRoom API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /auth/delete-account:
    delete:
      operationId: deleteAccount
      tags: [auth]
      summary: Delete current user account
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Account deleted
...
  /rooms/{roomId}/messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: List messages in a room (paginated)
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: before
          in: query
          description: Cursor - load messages before this message ID
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesResponse"
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      security:
        - bearerAuth: []
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{messageId}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Soft-delete a message (sender only)
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /messages/{messageId}/report:
    post:
      operationId: reportMessage
      tags: [messages]
      summary: Report a message
      security:
        - bearerAuth: []
      parameters:
        - name: messageId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportMessageRequest"
      responses:
        "201":
          description: Reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /blocks:
    get:
      operationId: listBlocks
      tags: [blocks]
      summary: List blocked users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Blocked users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedUser"
    post:
      operationId: blockUser
      tags: [blocks]
      summary: Block a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockUserRequest"
      responses:
        "201":
          description: Blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /blocks/{userId}:
    delete:
      operationId: unblockUser
      tags: [blocks]
      summary: Unblock a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Unblocked
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        username:
          type: string
          minLength: 3
          maxLength: 30
        avatarKey:
          type: string
      required:
        - email
        - password
        - username
        - avatarKey
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/Profile"
      required:
        - token
        - user
    Profile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        username:
          type: string
        avatarKey:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - username
        - avatarKey
        - createdAt
    UpdateProfileRequest:
      type: object
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        avatarKey:
          type: string
...
    Message:
      type: object
      properties:
        id:
          type: string
        roomId:
          type: string
        senderId:
          type: string
        senderUsername:
          type: string
        senderAvatarKey:
          type: string
        body:
          type: string
        replyToId:
          type: string
          nullable: true
        replyToBody:
          type: string
          nullable: true
        replyToSenderUsername:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        isDeleted:
          type: boolean
      required:
        - id
        - roomId
        - senderId
        - senderUsername
        - senderAvatarKey
        - body
        - createdAt
        - isDeleted
    MessagesResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message"
        hasMore:
          type: boolean
      required:
        - messages
        - hasMore
    SendMessageRequest:
      type: object
      properties:
        body:
          type: string
          maxLength: 2000
        replyToId:
          type: string
          nullable: true
      required:
        - body
    ReportMessageRequest:
      type: object
      properties:
        reason:
          type: string
          enum:
            - spam
            - harassment
            - hate
            - other
        note:
          type: string
          nullable: true
      required:
        - reason
    BlockUserRequest:
      type: object
      properties:
        blockedId:
          type: string
      required:
        - blockedId
    BlockedUser:
      type: object
      properties:
        userId:
          type: string
        username:
          type: string
        avatarKey:
          type: string
        blockedAt:
          type: string
          format: date-time
      required:
        - userId
        - username
        - avatarKey
        - blockedAt